% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{cdensplot}
\alias{cdensplot}
\title{cdensplot}
\usage{
cdensplot(cdens)
}
\arguments{
\item{x}{.csv file exported from Excel, see details below}
}
\value{
\describe{
\item{value}{ggplot showing results for each method for considered species}}
}
\description{
\code{cdensplot}  function dedicated for quick single-region data obs.
}
\details{
\usage{cdensplot(x)}

Here will be described how it works TBA
}

