% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compdens.R
\name{compdens}
\alias{compdens}
\title{compdens}
\usage{
compdens(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{data frame with mean, SE and results of ks tests for each method TBA}}
}
\description{
\code{compdens}  compute mean and SE for denisties using six methods, and p-values of ks tests
}
\details{
\usage{compdens(x, area)}

Here will be described how it works TBA
}

