% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compdens.R
\name{comptab}
\alias{comptab}
\title{comptab}
\usage{
comptab(x)
}
\arguments{
\item{x}{data frame containing mean, se and p-values of tests, computed by compdens}
}
\value{
\describe{
\item{value}{data frame with results for a specified region, sorted by species. TBA}}
}
\description{
\code{comptab}  compute mean and SE for denisties using six methods, and p-values of ks tests
}
\details{
\usage{comptab(x)}

Here will be described how it works TBA
}

