% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{dplotall}
\alias{dplotall}
\title{dplotall}
\usage{
dplotall(x, name)
}
\arguments{
\item{x}{data frame containing game densities, e.g. result of \code{game.load}}

\item{name}{name for .pdf file which will be saved}
}
\value{
\describe{
\item{value}{ggplot showing results for each method for considered species}}
}
\description{
\code{dplotall}  function dedicated for plotting all species
}
\details{
\usage{dplotall(x)}

Here will be described how it works TBA
}

