% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{game.load.pl}
\alias{game.load.pl}
\title{game.load.pl}
\usage{
game.load.pl(x, sep = ";")
}
\arguments{
\item{x}{.csv file exported from Excel in long format, see details below}

\item{sep}{separator of csv file, deflaut is ';'}
}
\value{
Data.frame without zero-obs species [in case of] densities within each plot
}
\description{
\code{game.load.pl}  loads .csv data from long table
}
\details{
\usage{game.load.pl(x, sep=';')}

To use file there is needed a specific file structure, with information about number of trail, trail area, and then - counts of each species. TBA
}

