% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{game.load}
\alias{game.load}
\title{game.load}
\usage{
game.load(x, sep)
}
\arguments{
\item{x}{.csv file exported from Excel in short format, see details below}

\item{sep}{separator of csv file, deflaut is ';'}
}
\value{
Data.frame with species densities within each plot
}
\description{
\code{game.load}  loads .csv data from long table
}
\details{
\usage{game.load(x, sep=';')}

To use file there is needed a file with an ordered structure - drive number, drive area and then - species in each column TBA
}

