% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liczymy.R
\name{liczymy}
\alias{liczymy}
\title{liczymy}
\usage{
liczymy(x, sep = ";", long = TRUE)
}
\arguments{
\item{x}{.csv file exported from Excel, see details below}

\item{sep}{separator of csv file, deflaut is ';'}

\item{long}{logical flag, whether table type is long or short}
}
\value{
\describe{
\item{value}{data frame with mean, SE and results of ks tests for each method TBA}}
}
\description{
\code{liczymy}  function dedicated for quick single-region data obs.
}
\details{
\usage{liczymy(x, sep=';', long=TRUE)}

Here will be described how it works TBA
}

