% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{mean.gam}
\alias{mean.gam}
\title{mean.gam}
\usage{
\method{mean}{gam}(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ estimated value computed from vector with Gamma distribution using \code{eGammaAlt} from \code{EnvStats} package}}
}
\description{
\code{mean.gam}  compute estimated value from vector of Gamma distribution
}
\details{
\usage{mean.gam(x, area)}

Function computes expected value and multiply by percent of area where species was observed. In case of n<2, weighted mean is used instead.Here will be described how it works TBA
}

