% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{mean.lnor}
\alias{mean.lnor}
\title{mean.lnor}
\usage{
\method{mean}{lnor}(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ estimated value computed from vector with lognormal distribution by... TBA}}
}
\description{
\code{mean.lnor}  compute estimated value from vector of lognormal distribution
}
\details{
\usage{mean.lnor(x, area)}

Function computes expected value and multiply by percent of area where species was observed. In case of n<2, weighted mean is used instead.Here will be described how it works TBA
}

