% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{mean.nor}
\alias{mean.nor}
\title{mean.nor}
\usage{
\method{mean}{nor}(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ estimated value computed from vector with normal distribution using \code{mean} from \code{base} package}}
}
\description{
\code{mean.nor}  compute estimated value from vector of normal distribution
}
\details{
\usage{mean.nor(x, area)}

Function computes expected value and multiply by percent of area where species was observed. In case of n<2, weighted mean is used instead.Here will be described how it works TBA
}

