% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{mean.rgeom}
\alias{mean.rgeom}
\title{mean.rgeom}
\usage{
\method{mean}{rgeom}(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ estimated value computed from vector with Geometric distribution using \code{fitdistr} from \code{MASS} package}}
}
\description{
\code{mean.rgeom}  compute estimated value from vector of Geometric distribution
}
\details{
\usage{mean.rgeom(x, area)}

Function computes expected value and multiply by percent of area where species was observed. In case of n<2, weighted mean is used instead.Here will be described how it works TBA
}

