% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{meangeo}
\alias{meangeo}
\title{meangeo}
\usage{
meangeo(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ geometric mean computed using \code{geoMean} from \code{EnvStats} package}}
}
\description{
\code{meangeo}  compute geometric mean
}
\details{
\usage{meangeo(x, area)}

Here will be described how it works TBA
}

