% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rysuj}
\alias{rysuj}
\title{rysuj}
\usage{
rysuj(x, spe = ";", long = TRUE)
}
\arguments{
\item{file}{name}
}
\value{
\describe{
\item{value}{ggplot showing results for each method for considered species}}
}
\description{
\code{rysuj}  polish wrapper for plotting all species
}
\details{
\usage{rysuj(x)}

Here will be described how it works TBA
}

