% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{se.geo}
\alias{se.geo}
\title{se.geo}
\usage{
se.geo(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ standard error computed as \code{sd/sqrt(n)} of non-zero observations, sd computed using \code{geoSD} from \code{EnvStats} package}}
}
\description{
\code{se.geo}  compute standard error off geometric mean
}
\details{
\usage{se.geo(x, area)}

Here will be described how it works TBA
}

