% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{se.lnor}
\alias{se.lnor}
\title{se.lnor}
\usage{
se.lnor(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ standard error computed as \code{sd/sqrt(n)} of non-zero observations, sd computed as... TBA}}
}
\description{
\code{se.lnor}  compute standard error from vector of lognormal distribution
}
\details{
\usage{se.lnor(x, area)}

Here will be described how it works TBA
}

