% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{se.nor}
\alias{se.nor}
\title{se.nor}
\usage{
se.nor(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ standard error computed as \code{sd/sqrt(n)} of non-zero observations}}
}
\description{
\code{se.nor}  compute standard error from vector of normal distribution
}
\details{
\usage{se.nor(x, area)}

Here will be described how it works TBA
}

