% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{se.ray}
\alias{se.ray}
\title{se.ray}
\usage{
se.ray(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ standard error computed as \code{sd/sqrt(n)} of non-zero observations, sd computed using \code{getRayParam} from \code{shotGroups} package}}
}
\description{
\code{se.ray}  compute standard error from vector of Rayleigh distribution
}
\details{
\usage{se.ray(x, area)}

Here will be described how it works TBA
}

