% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se.R
\name{se.rgeom}
\alias{se.rgeom}
\title{se.rgeom}
\usage{
se.rgeom(x, area)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}

\item{area}{a vector containing areas of each drives in ha}
}
\value{
\describe{
\item{value}{ standard error computed as \code{sd/sqrt(n)} of non-zero observations, sd computed using \code{fitdistr} from \code{MASS} package}}
}
\description{
\code{se.rgeom}  compute standard error from vector of geometric distribution
}
\details{
\usage{se.rgeom(x, area)}

Here will be described how it works TBA
}

