% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_tests.R
\name{ts.gam}
\alias{ts.gam}
\title{ts.gam}
\usage{
ts.gam(x)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}
}
\description{
\code{ts.gam}  tests empiric distribution versus Gamma theoretical distribution
}
\details{
\usage{ts.gam(x)}

Here will be described how it works TBA
}

