% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_tests.R
\name{ts.geom}
\alias{ts.geom}
\title{ts.geom}
\usage{
ts.geom(x)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}
}
\description{
\code{ts.geom}  tests empiric distribution versus Geometric theoretical distribution
}
\details{
\usage{ts.geom(x)}

Here will be described how it works TBA
}

