% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_tests.R
\name{ts.lnorm}
\alias{ts.lnorm}
\title{ts.lnorm}
\usage{
ts.lnorm(x)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}
}
\description{
\code{ts.lnorm}  tests empiric distribution versus lognormal theoretical distribution
}
\details{
\usage{ts.lnorm(x)}

Here will be described how it works TBA
}

