% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_tests.R
\name{ts.nor}
\alias{ts.nor}
\title{ts.nor}
\usage{
ts.nor(x)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}
}
\description{
function is a wrapper for KS normality test... TBA

\usage{ts.nor(x)}
}
\details{
\code{ts.nor}  tests empiric distribution versus Normal theoretical distribution

Here will be described how it works TBA
}

