% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_tests.R
\name{ts.ray}
\alias{ts.ray}
\title{ts.ray}
\usage{
ts.ray(x)
}
\arguments{
\item{x}{a vector containing observations, e.g. game densities}
}
\description{
\code{ts.ray}  tests empiric distribution versus Rayleigh theoretical distribution
}
\details{
\usage{ts.ray(x)}

Here will be described how it works TBA
}

